perf(frontend): lazy-load canvas-confetti + react-markdown to shrink initial bundle
Two heavy client-only modules now load on-demand instead of shipping in
every initial page bundle. Per the Next.js lazy-loading guide
(https://nextjs.org/docs/app/guides/lazy-loading) — "Loading External
Libraries" pattern for the fire-and-forget case, React.lazy + Suspense
for the renderable-component case.

1. canvas-confetti (~8 KB gzipped) — `lib/confetti.ts`
   Module-level promise cache calls `import("canvas-confetti")` lazily
   on first invocation; later calls reuse the cached promise. Public
   API (`fireBasic`, `fireRandomDirection`, `fireFireworks`,
   `fireSideCannons`, `fireOnUserAction`) stays synchronous void —
   call sites unchanged. Internal wrappers fire-and-forget the dynamic
   load with `void loadConfetti()?.then(confetti => ...)`. Returns
   null on the server (canvas-confetti requires `window`).

2. react-markdown + remark-gfm (~35-45 KB gzipped) —
   `components/chat/markdown-message.tsx` (new) +
   `components/chat/message-list.tsx`
   Extracted the markdown renderer + its 14-component Tailwind theme
   into its own module, loaded via `React.lazy`. The `<Suspense>`
   fallback renders the same raw text in a `whitespace-pre-wrap`
   paragraph — CLS stays at 0 because the unformatted text occupies
   identical vertical footprint to the formatted output for the
   ~50-150ms before the markdown chunk hydrates. First-time visitors
   on the locked landing page download zero markdown JS; only users
   who unlock the chat AND receive an assistant response trigger the
   download.

Why not lazy-load audio-orchestrator (deliberate skip):
   Chrome's autoplay policy requires the AudioContext to be
   constructed synchronously inside a user-gesture event handler.
   Dynamic-importing the orchestrator class would break that
   constraint (gesture frame ends before the chunk resolves), causing
   the autoplay-policy informational warning to appear in console —
   which Lighthouse counts as a Best Practices regression. Trading
   ~10-15 KB savings for a BP score regression is the wrong direction.

Test updates: two cases in `tests/message-list.test.tsx` switched
from sync `getByText` / `getByRole` to async `findByText` / `findByRole`
because the markdown DOM lands one microtask later than the synchronous
Suspense fallback in jsdom. The synchronous queries would race the
lazy load.

TypeScript: `canvas-confetti` ships CommonJS-style `export = confetti`
plus a `declare namespace`. `typeof import("canvas-confetti").default`
fails because the namespace has no `default` member. Resolved by
importing the static type (`import type confettiTypeImport from
"canvas-confetti"`) and resolving runtime via `mod.default ?? mod`
with an explicit cast — handles both module shapes.

Quality gates:
  - typecheck clean
  - biome lint clean
  - vitest: 172/172 pass
  - playwright a11y: 2/2 pass (locked + unlocked, zero WCAG violations)
  - production build clean
